Help gcc build when glibc is installed
When glibc is installed, programs need to be linked
with options that specify settings for the dynamic linker
and rpath, via the LDFLAGS environment variable.

The special compilation line used by one of the probes in
libgomp's autoconf generated configure script doesn't deal
with LDFLAGS appropriately, but [also] setting CFLAGS seems
to work.

See issue #257