lua: Bump revision for Linuxbrew
New revision needed due to readline SONAME change.
Also, change the bottle-pouring logic.

Note that the revision bump conditions on OS.mac? rather than on
OS.linux? even though the bottle-pouring logic change only applies
to Linux. This is because readline was a marked as a dependency
everywhere except macOS.