intltool: handle env that handles only 1 argument
On some ancient platforms, /usr/bin/env might be able to handle
only 1 argument. In such a case, 'perl -w' will be interpreted
as a single argument and 'env' will fail to execute it because
no file or script 'perl -w' exists on a normal platform. So, we
have to use a (better) `use warnings;` approach. When we do
`inreplace`, we exploit the fact that all scripts use strict module.