campari 3.0.1 (new formula)
This software has been already uploaded for launchpad and it is
fully GPL. It is made for analysing and simulating Molecular
Dynamics trajectories. We already compiled from source on MacOS
but we wanted to make the software available for a broader community.
The tests are still to complete, but we do extensive testing on gitlab
runners and the software is fully maintained since 2009 (its foundation).