explicitly specify Python.h path during config.
During config of libxslt, on some linux systems it cannot find the system's python header files, leading to failure in installation.

This commit points to python@2 installation managed by linuxbrew so the config script can find the header file. Python@2 is added as a dependency for the formula as well.