JAVA_HOME is set to MacOS home directory instead of Linux home directory
The root directory for Java installations is not in "Contents/Home/". This directory structure exists in MacOS but not in Linux. This formula does not work on Linux due to the incorrect environment variable, because Java cannot be found correctly. Instead, the opt_prefix directory should be used.

Based on https://docs.oracle.com/javase/10/install/installed-directory-structure-jdk-and-jre.htm and testing on Ubuntu.