xerces-c: remove non-relocatable block
In cc71aa7bd153fa145bf0e4df9ea4aeb943a85c0c (a long time ago)
we added this code to notify that the bottle was not relocatable.

Rebuild without that block to see happens. If it is not relocatable,
we will just remove the "cellar :any" line, as we do nowadays.