llvm: Re-enable and fix libc++ on Linux.
Fixes #13332 (again, it has broken and then been fixed a few times in
the interim from what I can tell).

This is essentially the combination of reverting #23059 and adding a fix
for the installation to avoid trying to use target-specific install
trees for runtimes on Linux. Without this, the runtimes end up in
a different structure than expected and flags like `-stdlib=libc=+`
don't work to find and use libc++ even though it is built and installed.

The original issue that #23059 claims to fix doesn't reproduce fro me on
an Ubuntu 16.04 machine, and the error message doesn't appear to have
anything to do with `libc++` and so I don't think there is anything
else to do here.

However, when the system GCC is sufficiently old, there were issues
relying on its runtime library `libgcc_s` for `libc++`. So this PR also
adds minimal use of LLVM's compiler builtin libraries (separately
`compiler-rt` and `libunwind`) to supplement this. Due to some
limitations in the LLVM CMake setup, this isn't enough to completely
avoid the dependency on the system `libgcc_s`, but should remove the
version constraint.

With this PR applied, LLVM-12 installs cleanly from source for me on
both a Debian testing machine and an Ubuntu 16.04 machine. Provided the
16.04 machine. All of `brew test llvm@12` stuff passes now. And Ubuntu
16.04 is the same baseline as the distribution in #23065 (sadly
I couldn't directly test Linux Mint as I couldn't find a VM with it
installed).

Note that this doesn't add any option for *using* libc++, it simply
builds and includes libc++ as part of the install LLVM and Clang
toolchain.

If anyone hits issues with LLVM builds in Homebrew, I'm happy to help
debug and help maintain the LLVM build here. I'm a very long time LLVM
developer and user with lots of experience with the build system.