llvm: fix (and test) static libc++ and LLVM runtimes
This adds the necessary CMake flags to pretty thoroughly enable linking
statically against libc++. With this, you can use `clang++` driver flags
to handle most aspects of the linking, and correctly get all the
relevant libraries. They also will reliably link shared objects. I've
added tests that both check the static linking can be used in position
independent code contexts and that check that none of `libc++`,
`libstdc++`, or `libgcc` end up linked when using these modes.

Details of the exact `clang++` flags needed are in the tests.

Some of the CMake flags here are really working around bugs:
https://bugs.llvm.org/show_bug.cgi?id=43604
https://bugs.llvm.org/show_bug.cgi?id=46321

For bug 43604, it seems unlikely to impact Homebrew unless the CMake
used for some reason changed its default, but it seemed worth fixing
this in place so that it works reliably.

For bug 46321, I've figured out the complex set of flags that cause
everything to work here. Hopefully the bug ends up with upstream
standardizing on this packaging approach and making the flags more
reasonable.